backend_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
fal-client==0.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.8.3
